TableViewCell 动态高度

+++ Woke/Interface/EWNotificationViewController.m
@@ -126,7 +126,7 @@
 
 - (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath{
     EWNotificationCell *cell = (EWNotificationCell*)[self tableView:_tableView cellForRowAtIndexPath:indexPath];
-    NSInteger h = cell.height;
+    CGFloat h = cell.height;
     return h;
 }
